requirements.txt
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
boto3==1.29.5
botocore==1.32.5
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.1.1
cssselect2==0.7.0
cssutils==2.9.0
cycler==0.10.0
Cython==3.0.8
datasets==2.14.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.25.1
dill==0.3.7
Django==4.2.7
dlib==19.24.2
dnspython==2.4.2
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.11
einops==0.7.0
ete3==3.1.3
exceptiongroup==1.2.0
executing==2.0.1
facelib @ git+https://github.com/sajjjadayobi/FaceLib.git@382841efec823f059f49754e5d378abb12cfb551
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.37
gradio==3.27.0
gradio_client==0.8.0
greenlet==3.0.1
-e git+https://github.com/IDEA-Research/GroundingDINO.git@57535c5a79791cb76e36fdb64975271354f10251#egg=groundingdino
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==2.10
imagecorruptions==1.1.2
imageio==2.31.5
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==6.8.0
insightface==0.7.3
iopath==0.1.10
ipdb==0.13.13
ipython
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==17.0.2
llvmlite==0.41.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
orderedset==2.0.3
orjson==3.9.10
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
Pillow==10.0.1
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.9.0
proglog==0.1.10
progressbar==2.5
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.9.7
pyarrow==14.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pymongo==4.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytorch-lightning==2.1.1
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.12
rpds-py==0.17.1
s3transfer==0.7.0
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.13.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
sentry-sdk==1.31.0
setproctitle==1.3.3
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
supervision==0.6.0
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.14.0
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-tools==0.1.5
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.6
uvicorn==0.25.0
visdom==0.2.4
wandb==0.15.11
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
widgetsnbextension==4.0.9
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
